+++ R/html_dependencies.R
@@ -221,13 +221,17 @@ html_dependency_navigation <- function(code_menu, source_embed) {
                  script = script)
 }
 
-html_dependency_anchor_sections <- function() {
-
+html_dependency_anchor_sections <- function(style = c("hash", "symbol", "icon"),
+                                            section_divs = FALSE) {
+  style <- match.arg(style)
+  # dynamically build css list
+  content_style <- function(style) sprintf("anchor-sections-%s.css", style)
   htmlDependency(name = "anchor-sections",
-                 version = "1.0.1",
+                 version = "1.1.0",
                  src = pkg_file("rmd/h/anchor-sections"),
-                 script = "anchor-sections.js",
-                 stylesheet = "anchor-sections.css")
+                 script = if (section_divs) "anchor-sections.js",
+                 stylesheet = c("anchor-sections.css", content_style(style))
+  )
 }
 
 # analyze navbar html source for icon dependencies