+++ tests/testthat/test-html_dependencies.R
@@ -1,6 +1,15 @@
 # TODO: to remove when switching the package to edition 3
 local_edition(3)
 
+test_that("Correct anchor_sections style is used", {
+  deps <- html_dependency_anchor_sections
+  expect_s3_class(deps(), "html_dependency")
+  expect_error(deps("dummy"), "should be one of")
+  expect_match(deps()$stylesheet[[2]], "anchor-sections-hash.css")
+  expect_null(deps()$script[[1]])
+  expect_equal(deps(section_divs = TRUE)$script[[1]], "anchor-sections.js")
+})
+
 test_that("dependency merge is correct", {
   # normalize lists for comparison by removing names and NULL elements
   prepare_list <- function(lst) {